Fixed go.mod files to fix CLI. Adding web.
