+++ go.mod
@@ -1,3 +1,7 @@
-module github.com/conflabermits/health_checker/hcfunc
+module github.com/conflabermits/health_checker
 
 go 1.19
+
+replace github.com/conflabermits/health_checker/hcfunc => ./pkg
+
+require github.com/conflabermits/health_checker/hcfunc v0.0.0-00010101000000-000000000000
