+++ pkg/go.mod
@@ -1,7 +1,5 @@
-module github.com/conflabermits/health_checker
+module github.com/conflabermits/health_checker/hcfunc
 
 go 1.19
 
 replace github.com/conflabermits/health_checker/hcfunc => ./pkg
-
-require github.com/conflabermits/health_checker/hcfunc v0.0.0-00010101000000-000000000000
